requirements.txt
ultralytics
streamlit
opencv-python
pyyaml
matplotlib
Pillow
